# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignedPatientDTO": {
        "properties": {
          "assigned_at": {
            "format": "date-time",
            "title": "Assigned At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_rural": {
            "default": false,
            "title": "Is Rural",
            "type": "boolean"
          },
          "lang_pref": {
            "title": "Lang Pref",
            "type": "string"
          }
        },
        "required": [
          "id",
          "lang_pref",
          "assigned_at"
        ],
        "title": "AssignedPatientDTO",
        "type": "object"
      },
...
      "Body_proxy_upload_api_v1_workers_patients__patient_id__documents_post": {
        "properties": {
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_proxy_upload_api_v1_workers_patients__patient_id__documents_post",
        "type": "object"
      },
...
      "CurrentTruthDTO": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/CurrentTruthEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          }
        },
        "required": [
          "patient_id",
          "entries",
          "generated_at"
        ],
        "title": "CurrentTruthDTO",
        "type": "object"
      },
      "CurrentTruthEntry": {
        "description": "One resolved fact about a patient. Pure reducer output (no DB id required).",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "entry_type": {
            "enum": [
              "medication",
              "lab_result",
              "diagnosis",
              "allergy",
              "vital"
            ],
            "title": "Entry Type",
            "type": "string"
          },
          "normalized_key": {
            "title": "Normalized Key",
            "type": "string"
          },
          "source_claim_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Source Claim Ids",
            "type": "array"
          },
          "state": {
            "default": "confirmed",
            "enum": [
              "confirmed",
              "needs_review",
              "conflict",
              "possibly_discontinued"
            ],
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "value": {
            "additionalProperties": true,
            "title": "Value",
            "type": "object"
          }
        },
        "required": [
          "entry_type",
          "normalized_key",
          "value",
          "confidence"
        ],
        "title": "CurrentTruthEntry",
        "type": "object"
      },
...
      "DocumentListItem": {
        "properties": {
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "source": {
            "default": "upload",
            "title": "Source",
            "type": "string"
          },
          "status": {
            "default": "pending",
            "title": "Status",
            "type": "string"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patient_id",
          "uploaded_at"
        ],
        "title": "DocumentListItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PatientCreateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "lang_pref": {
            "default": "mr",
            "title": "Lang Pref",
            "type": "string"
          }
        },
        "title": "PatientCreateRequest",
        "type": "object"
      },
      "PatientDTO": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lang_pref": {
            "default": "mr",
            "title": "Lang Pref",
            "type": "string"
          },
          "patient_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Token"
          }
        },
        "required": [
          "id",
          "created_at"
        ],
        "title": "PatientDTO",
        "type": "object"
      },
...
      "PatientUpdateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "lang_pref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang Pref"
          }
        },
        "title": "PatientUpdateRequest",
        "type": "object"
      },
...
      "ProxyPatientCreate": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_rural": {
            "default": false,
            "title": "Is Rural",
            "type": "boolean"
          },
          "lang_pref": {
            "default": "mr",
            "title": "Lang Pref",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "ProxyPatientCreate",
        "type": "object"
      },
...
      "ShareDTO": {
        "description": "Returned when a share is created — carries the link + QR for the patient.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "qr_svg": {
            "title": "Qr Svg",
            "type": "string"
          },
          "share_id": {
            "title": "Share Id",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "share_id",
          "token",
          "url",
          "qr_svg",
          "created_at"
        ],
        "title": "ShareDTO",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkerDTO": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "facility_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Name"
          },
          "facility_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Type"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at"
        ],
        "title": "WorkerDTO",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Supabase access token (required when SUPABASE_ENABLED=true)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SETU API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/patients": {
      "post": {
        "description": "Legacy anonymous patient create — disabled when Supabase auth is enabled.",
        "operationId": "create_patient_api_v1_patients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Patient",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/me": {
      "get": {
        "description": "Find or create the patient record linked to the authenticated Supabase user.",
        "operationId": "get_or_create_me_api_v1_patients_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDTO"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Or Create Me",
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "update_me_api_v1_patients_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "patients"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/documents": {
      "get": {
        "operationId": "list_patient_documents_api_v1_patients__patient_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentListItem"
                  },
                  "title": "Response List Patient Documents Api V1 Patients  Patient Id  Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Patient Documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/memory": {
      "get": {
        "operationId": "get_memory_api_v1_patients__patient_id__memory_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentTruthDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Memory",
        "tags": [
          "memory"
        ]
      }
    },
...
    "/api/v1/workers/me": {
      "get": {
        "operationId": "get_worker_me_api_v1_workers_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerDTO"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Worker Me",
        "tags": [
          "health-workers"
        ]
      }
    },
    "/api/v1/workers/patients": {
      "get": {
        "operationId": "list_assigned_patients_api_v1_workers_patients_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssignedPatientDTO"
                  },
                  "title": "Response List Assigned Patients Api V1 Workers Patients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Assigned Patients",
        "tags": [
          "health-workers"
        ]
      },
      "post": {
        "operationId": "register_patient_api_v1_workers_patients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxyPatientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Patient",
        "tags": [
          "health-workers"
        ]
      }
    },
    "/api/v1/workers/patients/{patient_id}/documents": {
      "post": {
        "operationId": "proxy_upload_api_v1_workers_patients__patient_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_proxy_upload_api_v1_workers_patients__patient_id__documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Proxy Upload Api V1 Workers Patients  Patient Id  Documents Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Upload",
        "tags": [
          "health-workers"
        ]
      }
    },
    "/api/v1/workers/patients/{patient_id}/share": {
      "post": {
        "operationId": "proxy_share_api_v1_workers_patients__patient_id__share_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Share",
        "tags": [
          "health-workers"
        ]
      }
    },
...
  }
}